Create main.cpp
Write an application that models a bank profile containing a savings account and a bank card
associate. We consider that only one operation can be performed per month (except profile query) and that
the interest related to the savings account, expressed as a percentage, is applied monthly, only on the balance related to the account
of savings. The following information is read from the keyboard, on different lines: a. a string of characters representing the name of the account holder (it may contain more than one word); b. a real value representing the card balance; c. a real value representing the balance of the savings account; d. a real value representing the savings account interest;